+++ R/MultipleGroup-methods.R
@@ -250,16 +250,3 @@ setMethod(
         ret
     }
 )
-
-setMethod(
-    f = "fitted",
-    signature = signature(object = 'MultipleGroupClass'),
-    definition = function(object, ...)
-    {
-        ret <- vector('list', length(object@groupNames))
-        names(ret) <- object@groupNames
-        for(g in 1:length(ret))
-            ret[[g]] <- fitted(object@cmods[[g]], ...)
-        ret
-    }
-)